Consolidate the various register traits.
RegistersStore and RegistersLoad are consolidated into the new
trait RegistersTrait. RegistersSigner now has RegistersTrait
as a supertrait, so that functions accepting it don't need
to write &(impl RegistersSigner<N> + RegistersTrait<N>).

Similarly FinalizeRegistersTrait also has RegistersTrait as
a supertrait.

RegistersStoreCircuit, RegistersLoadCircuit, and RegistersSignerCircuit
have now been consolidated into a new trait, RegistersCircuit.
(Registers is the only type implementing this trait, but as with
the stack traits it exists so that `synthesizer-program` can
use it without needing to have a concrete implementation.)

Some uses of these traits in `synthesizer-process` now refer to the
concrete Registers type.

The `traits` module in `synthesizer-program` is now private,
and `use`s of it now pull traits directly from the crate.

The `RegistersCall` trait served no purpose as a separate
trait and its function is now directly in the Registers type.

Some functions in `synthesizer_process` now accept a concrete
`Registers` type.
