+++ synthesizer/process/src/stack/registers/caller.rs
@@ -64,53 +64,3 @@ impl<N: Network, A: circuit::Aleo<Network = N>> RegistersSigner<N> for Registers
         self.tvk = Some(tvk);
     }
 }
-
-impl<N: Network, A: circuit::Aleo<Network = N>> RegistersSignerCircuit<N, A> for Registers<N, A> {
-    /// Returns the transition signer, as a circuit.
-    #[inline]
-    fn signer_circuit(&self) -> Result<circuit::Address<A>> {
-        self.signer_circuit.clone().ok_or_else(|| anyhow!("Signer address (circuit) is not set in the registers."))
-    }
-
-    /// Sets the transition signer, as a circuit.
-    #[inline]
-    fn set_signer_circuit(&mut self, signer_circuit: circuit::Address<A>) {
-        self.signer_circuit = Some(signer_circuit);
-    }
-
-    /// Returns the root transition view key, as a circuit.
-    #[inline]
-    fn root_tvk_circuit(&self) -> Result<circuit::Field<A>> {
-        self.root_tvk_circuit.clone().ok_or_else(|| anyhow!("Root tvk (circuit) is not set in the registers."))
-    }
-
-    /// Sets the root transition view key, as a circuit.
-    #[inline]
-    fn set_root_tvk_circuit(&mut self, root_tvk_circuit: circuit::Field<A>) {
-        self.root_tvk_circuit = Some(root_tvk_circuit);
-    }
-
-    /// Returns the transition caller, as a circuit.
-    #[inline]
-    fn caller_circuit(&self) -> Result<circuit::Address<A>> {
-        self.caller_circuit.clone().ok_or_else(|| anyhow!("Caller address (circuit) is not set in the registers."))
-    }
-
-    /// Sets the transition caller, as a circuit.
-    #[inline]
-    fn set_caller_circuit(&mut self, caller_circuit: circuit::Address<A>) {
-        self.caller_circuit = Some(caller_circuit);
-    }
-
-    /// Returns the transition view key, as a circuit.
-    #[inline]
-    fn tvk_circuit(&self) -> Result<circuit::Field<A>> {
-        self.tvk_circuit.clone().ok_or_else(|| anyhow!("Transition view key (circuit) is not set in the registers."))
-    }
-
-    /// Sets the transition view key, as a circuit.
-    #[inline]
-    fn set_tvk_circuit(&mut self, tvk_circuit: circuit::Field<A>) {
-        self.tvk_circuit = Some(tvk_circuit);
-    }
-}
