+++ synthesizer/process/src/stack/registers/mod.rs
@@ -13,27 +13,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod call;
-mod caller;
-mod load;
-mod store;
+mod registers_circuit;
+mod registers_trait;
 
-use crate::{CallStack, RegisterTypes, RegistersCall};
+use crate::{CallStack, RegisterTypes};
 use console::{
     network::prelude::*,
     program::{Entry, Literal, Plaintext, Register, Value},
     types::{Address, Field},
 };
-use synthesizer_program::{
-    Operand,
-    RegistersLoad,
-    RegistersLoadCircuit,
-    RegistersSigner,
-    RegistersSignerCircuit,
-    RegistersStore,
-    RegistersStoreCircuit,
-    StackTrait,
-};
+use synthesizer_program::{Operand, RegistersCircuit, RegistersSigner, RegistersTrait, StackTrait};
 
 use indexmap::IndexMap;
 
@@ -66,6 +55,12 @@ pub struct Registers<N: Network, A: circuit::Aleo<Network = N>> {
 }
 
 impl<N: Network, A: circuit::Aleo<Network = N>> Registers<N, A> {
+    /// Returns the current call stack.
+    #[inline]
+    pub fn call_stack(&self) -> CallStack<N> {
+        self.call_stack.clone()
+    }
+
     /// Initializes a new set of registers, given the call stack.
     #[inline]
     pub fn new(call_stack: CallStack<N>, register_types: RegisterTypes<N>) -> Self {
