+++ synthesizer/process/src/stack/registers/registers_circuit.rs
@@ -15,78 +15,60 @@
 
 use super::*;
 
-impl<N: Network, A: circuit::Aleo<Network = N>> RegistersLoad<N> for Registers<N, A> {
-    /// Loads the value of a given operand from the registers.
-    ///
-    /// # Errors
-    /// This method will halt if the register locator is not found.
-    /// In the case of register accesses, this method will halt if the access is not found.
+impl<N: Network, A: circuit::Aleo<Network = N>> RegistersCircuit<N, A> for Registers<N, A> {
+    /// Returns the transition signer, as a circuit.
     #[inline]
-    fn load(&self, stack: &impl StackTrait<N>, operand: &Operand<N>) -> Result<Value<N>> {
-        // Retrieve the register.
-        let register = match operand {
-            // If the operand is a literal, return the literal.
-            Operand::Literal(literal) => return Ok(Value::Plaintext(Plaintext::from(literal))),
-            // If the operand is a register, load the value from the register.
-            Operand::Register(register) => register,
-            // If the operand is the program ID, load the program address.
-            Operand::ProgramID(program_id) => {
-                return Ok(Value::Plaintext(Plaintext::from(Literal::Address(program_id.to_address()?))));
-            }
-            // If the operand is the signer, load the value of the signer.
-            Operand::Signer => return Ok(Value::Plaintext(Plaintext::from(Literal::Address(self.signer()?)))),
-            // If the operand is the caller, load the value of the caller.
-            Operand::Caller => return Ok(Value::Plaintext(Plaintext::from(Literal::Address(self.caller()?)))),
-            // If the operand is the block height, throw an error.
-            Operand::BlockHeight => bail!("Cannot load the block height in a non-finalize context"),
-            // If the operand is the network ID, throw an error.
-            Operand::NetworkID => bail!("Cannot load the network ID in a non-finalize context"),
-        };
-
-        // Retrieve the stack value.
-        let stack_value =
-            self.console_registers.get(&register.locator()).ok_or_else(|| anyhow!("'{register}' does not exist"))?;
-
-        // Return the value for the given register or register access.
-        let stack_value = match register {
-            // If the register is a locator, then return the stack value.
-            Register::Locator(..) => stack_value.clone(),
-            // If the register is a register access, then load the specific stack value.
-            Register::Access(_, ref path) => {
-                match stack_value {
-                    // Retrieve the plaintext member from the path.
-                    Value::Plaintext(plaintext) => Value::Plaintext(plaintext.find(path)?),
-                    // Retrieve the record entry from the path.
-                    Value::Record(record) => match record.find(path)? {
-                        Entry::Constant(plaintext) | Entry::Public(plaintext) | Entry::Private(plaintext) => {
-                            Value::Plaintext(plaintext)
-                        }
-                    },
-                    // Retrieve the argument from the future.
-                    Value::Future(future) => future.find(path)?,
-                }
-            }
-        };
-
-        // Retrieve the register type.
-        match self.register_types.get_type(stack, register) {
-            // Ensure the stack value matches the register type.
-            Ok(register_type) => stack.matches_register_type(&stack_value, &register_type)?,
-            // Ensure the register is defined.
-            Err(error) => bail!("Register '{register}' is not a member of the function: {error}"),
-        };
-
-        Ok(stack_value)
+    fn signer_circuit(&self) -> Result<circuit::Address<A>> {
+        self.signer_circuit.clone().ok_or_else(|| anyhow!("Signer address (circuit) is not set in the registers."))
     }
-}
 
-impl<N: Network, A: circuit::Aleo<Network = N>> RegistersLoadCircuit<N, A> for Registers<N, A> {
+    /// Sets the transition signer, as a circuit.
+    #[inline]
+    fn set_signer_circuit(&mut self, signer_circuit: circuit::Address<A>) {
+        self.signer_circuit = Some(signer_circuit);
+    }
+
+    /// Returns the root transition view key, as a circuit.
+    #[inline]
+    fn root_tvk_circuit(&self) -> Result<circuit::Field<A>> {
+        self.root_tvk_circuit.clone().ok_or_else(|| anyhow!("Root tvk (circuit) is not set in the registers."))
+    }
+
+    /// Sets the root transition view key, as a circuit.
+    #[inline]
+    fn set_root_tvk_circuit(&mut self, root_tvk_circuit: circuit::Field<A>) {
+        self.root_tvk_circuit = Some(root_tvk_circuit);
+    }
+
+    /// Returns the transition caller, as a circuit.
+    #[inline]
+    fn caller_circuit(&self) -> Result<circuit::Address<A>> {
+        self.caller_circuit.clone().ok_or_else(|| anyhow!("Caller address (circuit) is not set in the registers."))
+    }
+
+    /// Sets the transition caller, as a circuit.
+    #[inline]
+    fn set_caller_circuit(&mut self, caller_circuit: circuit::Address<A>) {
+        self.caller_circuit = Some(caller_circuit);
+    }
+
+    /// Returns the transition view key, as a circuit.
+    #[inline]
+    fn tvk_circuit(&self) -> Result<circuit::Field<A>> {
+        self.tvk_circuit.clone().ok_or_else(|| anyhow!("Transition view key (circuit) is not set in the registers."))
+    }
+
+    /// Sets the transition view key, as a circuit.
+    #[inline]
+    fn set_tvk_circuit(&mut self, tvk_circuit: circuit::Field<A>) {
+        self.tvk_circuit = Some(tvk_circuit);
+    }
+
     /// Loads the value of a given operand from the registers.
     ///
     /// # Errors
     /// This method will halt if the register locator is not found.
     /// In the case of register accesses, this method will halt if the access is not found.
-    #[inline]
     fn load_circuit(&self, stack: &impl StackTrait<N>, operand: &Operand<N>) -> Result<circuit::Value<A>> {
         use circuit::Inject;
 
@@ -163,5 +145,51 @@ impl<N: Network, A: circuit::Aleo<Network = N>> RegistersLoadCircuit<N, A> for R
         };
 
         Ok(circuit_value)
+    }
+
+    /// Assigns the given value to the given register, assuming the register is not already assigned.
+    ///
+    /// # Errors
+    /// This method will halt if the given register is a register access.
+    /// This method will halt if the given register is an input register.
+    /// This method will halt if the register is already used.
+    fn store_circuit(
+        &mut self,
+        stack: &impl StackTrait<N>,
+        register: &Register<N>,
+        circuit_value: circuit::Value<A>,
+    ) -> Result<()> {
+        match register {
+            Register::Locator(locator) => {
+                // Ensure the register assignments are monotonically increasing.
+                let expected_locator = self.circuit_registers.len() as u64;
+                ensure!(expected_locator == *locator, "Out-of-order write operation at '{register}'");
+                // Ensure the register does not already exist.
+                ensure!(
+                    !self.circuit_registers.contains_key(locator),
+                    "Cannot write to occupied register '{register}'"
+                );
+
+                // Ensure the register type is valid.
+                match self.register_types.get_type(stack, register) {
+                    // Ensure the stack value matches the register type.
+                    Ok(register_type) => {
+                        stack.matches_register_type(&circuit::Eject::eject_value(&circuit_value), &register_type)?
+                    }
+                    // Ensure the register is defined.
+                    Err(error) => bail!("Register '{register}' is missing a type definition: {error}"),
+                };
+
+                // Store the stack value.
+                match self.circuit_registers.insert(*locator, circuit_value) {
+                    // Ensure the register has not been previously stored.
+                    Some(..) => bail!("Attempted to write to register '{register}' again"),
+                    // Return on success.
+                    None => Ok(()),
+                }
+            }
+            // Ensure the register is not a register access.
+            Register::Access(..) => bail!("Cannot store to a register access: '{register}'"),
+        }
     }
 }
