--- synthesizer/process/src/stack/registers/store.rs
@@ -1,107 +1,0 @@
-// Copyright (c) 2019-2025 Provable Inc.
-// This file is part of the snarkVM library.
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at:
-
-// http://www.apache.org/licenses/LICENSE-2.0
-
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use super::*;
-
-impl<N: Network, A: circuit::Aleo<Network = N>> RegistersStore<N> for Registers<N, A> {
-    /// Assigns the given value to the given register, assuming the register is not already assigned.
-    ///
-    /// # Errors
-    /// This method will halt if the given register is a register access.
-    /// This method will halt if the given register is an input register.
-    /// This method will halt if the register is already used.
-    #[inline]
-    fn store(&mut self, stack: &impl StackTrait<N>, register: &Register<N>, stack_value: Value<N>) -> Result<()> {
-        match register {
-            Register::Locator(locator) => {
-                // Ensure the register assignments are monotonically increasing.
-                let expected_locator = self.console_registers.len() as u64;
-                ensure!(expected_locator == *locator, "Out-of-order write operation at '{register}'");
-                // Ensure the register does not already exist.
-                ensure!(
-                    !self.console_registers.contains_key(locator),
-                    "Cannot write to occupied register '{register}'"
-                );
-
-                // Ensure the register type is valid.
-                match self.register_types.get_type(stack, register) {
-                    // Ensure the stack value matches the register type.
-                    Ok(register_type) => stack.matches_register_type(&stack_value, &register_type)?,
-                    // Ensure the register is defined.
-                    Err(error) => bail!("Register '{register}' is missing a type definition: {error}"),
-                };
-
-                // Store the stack value.
-                match self.console_registers.insert(*locator, stack_value) {
-                    // Ensure the register has not been previously stored.
-                    Some(..) => bail!("Attempted to write to register '{register}' again"),
-                    // Return on success.
-                    None => Ok(()),
-                }
-            }
-            // Ensure the register is not a register access.
-            Register::Access(..) => bail!("Cannot store to a register access: '{register}'"),
-        }
-    }
-}
-
-impl<N: Network, A: circuit::Aleo<Network = N>> RegistersStoreCircuit<N, A> for Registers<N, A> {
-    /// Assigns the given value to the given register, assuming the register is not already assigned.
-    ///
-    /// # Errors
-    /// This method will halt if the given register is a register access.
-    /// This method will halt if the given register is an input register.
-    /// This method will halt if the register is already used.
-    #[inline]
-    fn store_circuit(
-        &mut self,
-        stack: &impl StackTrait<N>,
-        register: &Register<N>,
-        circuit_value: circuit::Value<A>,
-    ) -> Result<()> {
-        match register {
-            Register::Locator(locator) => {
-                // Ensure the register assignments are monotonically increasing.
-                let expected_locator = self.circuit_registers.len() as u64;
-                ensure!(expected_locator == *locator, "Out-of-order write operation at '{register}'");
-                // Ensure the register does not already exist.
-                ensure!(
-                    !self.circuit_registers.contains_key(locator),
-                    "Cannot write to occupied register '{register}'"
-                );
-
-                // Ensure the register type is valid.
-                match self.register_types.get_type(stack, register) {
-                    // Ensure the stack value matches the register type.
-                    Ok(register_type) => {
-                        stack.matches_register_type(&circuit::Eject::eject_value(&circuit_value), &register_type)?
-                    }
-                    // Ensure the register is defined.
-                    Err(error) => bail!("Register '{register}' is missing a type definition: {error}"),
-                };
-
-                // Store the stack value.
-                match self.circuit_registers.insert(*locator, circuit_value) {
-                    // Ensure the register has not been previously stored.
-                    Some(..) => bail!("Attempted to write to register '{register}' again"),
-                    // Return on success.
-                    None => Ok(()),
-                }
-            }
-            // Ensure the register is not a register access.
-            Register::Access(..) => bail!("Cannot store to a register access: '{register}'"),
-        }
-    }
-}
