+++ synthesizer/program/src/logic/command/get_or_use.rs
@@ -13,12 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{
-    CallOperator,
-    Opcode,
-    Operand,
-    traits::{FinalizeStoreTrait, RegistersLoad, RegistersStore, StackTrait},
-};
+use crate::{CallOperator, FinalizeStoreTrait, Opcode, Operand, RegistersTrait, StackTrait};
 use console::{
     network::prelude::*,
     program::{Register, Value},
@@ -106,7 +101,7 @@ impl<N: Network> GetOrUse<N> {
         &self,
         stack: &impl StackTrait<N>,
         store: &impl FinalizeStoreTrait<N>,
-        registers: &mut (impl RegistersLoad<N> + RegistersStore<N>),
+        registers: &mut impl RegistersTrait<N>,
     ) -> Result<()> {
         // Determine the program ID and mapping name.
         let (program_id, mapping_name) = match self.mapping {
