+++ synthesizer/program/src/logic/command/rand_chacha.rs
@@ -13,12 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{
-    FinalizeRegistersState,
-    Opcode,
-    Operand,
-    traits::{RegistersLoad, RegistersStore, StackTrait},
-};
+use crate::{FinalizeRegistersState, Opcode, Operand, StackTrait};
 use console::{
     network::prelude::*,
     program::{Literal, LiteralType, Plaintext, Register, Value},
@@ -75,11 +70,7 @@ impl<N: Network> RandChaCha<N> {
 impl<N: Network> RandChaCha<N> {
     /// Finalizes the command.
     #[inline]
-    pub fn finalize(
-        &self,
-        stack: &impl StackTrait<N>,
-        registers: &mut (impl RegistersLoad<N> + RegistersStore<N> + FinalizeRegistersState<N>),
-    ) -> Result<()> {
+    pub fn finalize(&self, stack: &impl StackTrait<N>, registers: &mut impl FinalizeRegistersState<N>) -> Result<()> {
         // Ensure the number of operands is within bounds.
         if self.operands.len() > MAX_ADDITIONAL_SEEDS {
             bail!("The number of operands must be <= {MAX_ADDITIONAL_SEEDS}")
