+++ synthesizer/program/src/logic/instruction/operation/hash.rs
@@ -13,11 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{
-    Opcode,
-    Operand,
-    traits::{RegistersLoad, RegistersLoadCircuit, RegistersStore, RegistersStoreCircuit, StackTrait},
-};
+use crate::{Opcode, Operand, RegistersCircuit, RegistersTrait, StackTrait};
 use console::{
     network::prelude::*,
     program::{Literal, LiteralType, Plaintext, PlaintextType, Register, RegisterType, Value},
@@ -129,7 +125,6 @@ pub struct HashInstruction<N: Network, const VARIANT: u8> {
 
 impl<N: Network, const VARIANT: u8> HashInstruction<N, VARIANT> {
     /// Initializes a new `hash` instruction.
-    #[inline]
     pub fn new(
         operands: Vec<Operand<N>>,
         destination: Register<N>,
@@ -146,7 +141,6 @@ impl<N: Network, const VARIANT: u8> HashInstruction<N, VARIANT> {
     }
 
     /// Returns the opcode.
-    #[inline]
     pub const fn opcode() -> Opcode {
         match VARIANT {
             0 => Opcode::Hash("hash.bhp256"),
@@ -172,7 +166,6 @@ impl<N: Network, const VARIANT: u8> HashInstruction<N, VARIANT> {
     }
 
     /// Returns the operands in the operation.
-    #[inline]
     pub fn operands(&self) -> &[Operand<N>] {
         // Sanity check that the operands is the correct length.
         debug_assert!(
@@ -265,12 +258,7 @@ fn evaluate_hash_internal<N: Network>(
 
 impl<N: Network, const VARIANT: u8> HashInstruction<N, VARIANT> {
     /// Evaluates the instruction.
-    #[inline]
-    pub fn evaluate(
-        &self,
-        stack: &impl StackTrait<N>,
-        registers: &mut (impl RegistersLoad<N> + RegistersStore<N>),
-    ) -> Result<()> {
+    pub fn evaluate(&self, stack: &impl StackTrait<N>, registers: &mut impl RegistersTrait<N>) -> Result<()> {
         // Ensure the number of operands is correct.
         check_number_of_operands(VARIANT, Self::opcode(), self.operands.len())?;
         // Ensure the destination type is valid.
@@ -286,11 +274,10 @@ impl<N: Network, const VARIANT: u8> HashInstruction<N, VARIANT> {
     }
 
     /// Executes the instruction.
-    #[inline]
     pub fn execute<A: circuit::Aleo<Network = N>>(
         &self,
         stack: &impl StackTrait<N>,
-        registers: &mut (impl RegistersLoadCircuit<N, A> + RegistersStoreCircuit<N, A>),
+        registers: &mut impl RegistersCircuit<N, A>,
     ) -> Result<()> {
         use circuit::traits::{ToBits, ToFields};
 
@@ -312,16 +299,11 @@ impl<N: Network, const VARIANT: u8> HashInstruction<N, VARIANT> {
 
     /// Finalizes the instruction.
     #[inline]
-    pub fn finalize(
-        &self,
-        stack: &impl StackTrait<N>,
-        registers: &mut (impl RegistersLoad<N> + RegistersStore<N>),
-    ) -> Result<()> {
+    pub fn finalize(&self, stack: &impl StackTrait<N>, registers: &mut impl RegistersTrait<N>) -> Result<()> {
         self.evaluate(stack, registers)
     }
 
     /// Returns the output type from the given program and input types.
-    #[inline]
     pub fn output_types(
         &self,
         _stack: &impl StackTrait<N>,
@@ -346,7 +328,6 @@ impl<N: Network, const VARIANT: u8> HashInstruction<N, VARIANT> {
 
 impl<N: Network, const VARIANT: u8> Parser for HashInstruction<N, VARIANT> {
     /// Parses a string into an operation.
-    #[inline]
     fn parse(string: &str) -> ParserResult<Self> {
         /// Parse the operands from the string.
         fn parse_operands<N: Network>(string: &str, num_operands: usize) -> ParserResult<Vec<Operand<N>>> {
@@ -403,7 +384,6 @@ impl<N: Network, const VARIANT: u8> FromStr for HashInstruction<N, VARIANT> {
     type Err = Error;
 
     /// Parses a string into an operation.
-    #[inline]
     fn from_str(string: &str) -> Result<Self> {
         match Self::parse(string) {
             Ok((remainder, object)) => {
