+++ synthesizer/program/src/logic/instruction/operation/is.rs
@@ -13,11 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::{
-    Opcode,
-    Operand,
-    traits::{RegistersLoad, RegistersLoadCircuit, RegistersStore, RegistersStoreCircuit, StackTrait},
-};
+use crate::{Opcode, Operand, RegistersCircuit, RegistersTrait, StackTrait};
 use console::{
     network::prelude::*,
     program::{Literal, LiteralType, Plaintext, PlaintextType, Register, RegisterType, Value},
@@ -45,7 +41,6 @@ pub struct IsInstruction<N: Network, const VARIANT: u8> {
 
 impl<N: Network, const VARIANT: u8> IsInstruction<N, VARIANT> {
     /// Initializes a new `is` instruction.
-    #[inline]
     pub fn new(operands: Vec<Operand<N>>, destination: Register<N>) -> Result<Self> {
         // Sanity check the number of operands.
         ensure!(operands.len() == 2, "Instruction '{}' must have two operands", Self::opcode());
@@ -54,7 +49,6 @@ impl<N: Network, const VARIANT: u8> IsInstruction<N, VARIANT> {
     }
 
     /// Returns the opcode.
-    #[inline]
     pub const fn opcode() -> Opcode {
         match VARIANT {
             0 => Opcode::Is("is.eq"),
@@ -81,12 +75,7 @@ impl<N: Network, const VARIANT: u8> IsInstruction<N, VARIANT> {
 
 impl<N: Network, const VARIANT: u8> IsInstruction<N, VARIANT> {
     /// Evaluates the instruction.
-    #[inline]
-    pub fn evaluate(
-        &self,
-        stack: &impl StackTrait<N>,
-        registers: &mut (impl RegistersLoad<N> + RegistersStore<N>),
-    ) -> Result<()> {
+    pub fn evaluate(&self, stack: &impl StackTrait<N>, registers: &mut impl RegistersTrait<N>) -> Result<()> {
         // Ensure the number of operands is correct.
         if self.operands.len() != 2 {
             bail!("Instruction '{}' expects 2 operands, found {} operands", Self::opcode(), self.operands.len())
@@ -107,11 +96,10 @@ impl<N: Network, const VARIANT: u8> IsInstruction<N, VARIANT> {
     }
 
     /// Executes the instruction.
-    #[inline]
     pub fn execute<A: circuit::Aleo<Network = N>>(
         &self,
         stack: &impl StackTrait<N>,
-        registers: &mut (impl RegistersLoadCircuit<N, A> + RegistersStoreCircuit<N, A>),
+        registers: &mut impl RegistersCircuit<N, A>,
     ) -> Result<()> {
         // Ensure the number of operands is correct.
         if self.operands.len() != 2 {
@@ -136,16 +124,11 @@ impl<N: Network, const VARIANT: u8> IsInstruction<N, VARIANT> {
 
     /// Finalizes the instruction.
     #[inline]
-    pub fn finalize(
-        &self,
-        stack: &impl StackTrait<N>,
-        registers: &mut (impl RegistersLoad<N> + RegistersStore<N>),
-    ) -> Result<()> {
+    pub fn finalize(&self, stack: &impl StackTrait<N>, registers: &mut impl RegistersTrait<N>) -> Result<()> {
         self.evaluate(stack, registers)
     }
 
     /// Returns the output type from the given program and input types.
-    #[inline]
     pub fn output_types(
         &self,
         _stack: &impl StackTrait<N>,
@@ -178,7 +161,6 @@ impl<N: Network, const VARIANT: u8> IsInstruction<N, VARIANT> {
 
 impl<N: Network, const VARIANT: u8> Parser for IsInstruction<N, VARIANT> {
     /// Parses a string into an operation.
-    #[inline]
     fn parse(string: &str) -> ParserResult<Self> {
         // Parse the opcode from the string.
         let (string, _) = tag(*Self::opcode())(string)?;
@@ -207,7 +189,6 @@ impl<N: Network, const VARIANT: u8> FromStr for IsInstruction<N, VARIANT> {
     type Err = Error;
 
     /// Parses a string into an operation.
-    #[inline]
     fn from_str(string: &str) -> Result<Self> {
         match Self::parse(string) {
             Ok((remainder, object)) => {
