+++ synthesizer/program/src/traits/stack_and_registers.rs
@@ -39,6 +39,9 @@ use console::{
 use rand::{CryptoRng, Rng};
 use synthesizer_snark::{ProvingKey, VerifyingKey};
 
+/// This trait is intended to be implemented only by `synthesizer_process::Stack`.
+///
+/// We make it a trait only to avoid circular dependencies.
 pub trait StackTrait<N: Network> {
     /// Returns `true` if the proving key for the given function name exists.
     fn contains_proving_key(&self, function_name: &Identifier<N>) -> bool;
@@ -134,7 +137,7 @@ pub trait StackTrait<N: Network> {
     ) -> Result<Record<N, Plaintext<N>>>;
 }
 
-pub trait FinalizeRegistersState<N: Network> {
+pub trait FinalizeRegistersState<N: Network>: RegistersTrait<N> {
     /// Returns the global state for the finalize scope.
     fn state(&self) -> &FinalizeGlobalState;
 
@@ -148,7 +151,7 @@ pub trait FinalizeRegistersState<N: Network> {
     fn nonce(&self) -> u64;
 }
 
-pub trait RegistersSigner<N: Network> {
+pub trait RegistersSigner<N: Network>: RegistersTrait<N> {
     /// Returns the transition signer.
     fn signer(&self) -> Result<Address<N>>;
 
@@ -174,7 +177,68 @@ pub trait RegistersSigner<N: Network> {
     fn set_tvk(&mut self, tvk: Field<N>);
 }
 
-pub trait RegistersSignerCircuit<N: Network, A: circuit::Aleo<Network = N>> {
+pub trait RegistersTrait<N: Network> {
+    /// Loads the value of a given operand.
+    ///
+    /// # Errors
+    /// This method should halt if the register locator is not found.
+    /// In the case of register members, this method should halt if the member is not found.
+    fn load(&self, stack: &impl StackTrait<N>, operand: &Operand<N>) -> Result<Value<N>>;
+
+    /// Loads the literal of a given operand.
+    ///
+    /// # Errors
+    /// This method should halt if the given operand is not a literal.
+    /// This method should halt if the register locator is not found.
+    /// In the case of register members, this method should halt if the member is not found.
+    fn load_literal(&self, stack: &impl StackTrait<N>, operand: &Operand<N>) -> Result<Literal<N>> {
+        match self.load(stack, operand)? {
+            Value::Plaintext(Plaintext::Literal(literal, ..)) => Ok(literal),
+            Value::Plaintext(Plaintext::Struct(..))
+            | Value::Plaintext(Plaintext::Array(..))
+            | Value::Record(..)
+            | Value::Future(..) => {
+                bail!("Operand must be a literal")
+            }
+        }
+    }
+
+    /// Loads the plaintext of a given operand.
+    ///
+    /// # Errors
+    /// This method should halt if the given operand is not a plaintext.
+    /// This method should halt if the register locator is not found.
+    /// In the case of register members, this method should halt if the member is not found.
+    fn load_plaintext(&self, stack: &impl StackTrait<N>, operand: &Operand<N>) -> Result<Plaintext<N>> {
+        match self.load(stack, operand)? {
+            Value::Plaintext(plaintext) => Ok(plaintext),
+            Value::Record(..) | Value::Future(..) => bail!("Operand must be a plaintext"),
+        }
+    }
+
+    /// Assigns the given value to the given register, assuming the register is not already assigned.
+    ///
+    /// # Errors
+    /// This method should halt if the given register is a register member.
+    /// This method should halt if the given register is an input register.
+    /// This method should halt if the register is already used.
+    fn store(&mut self, stack: &impl StackTrait<N>, register: &Register<N>, stack_value: Value<N>) -> Result<()>;
+
+    /// Assigns the given literal to the given register, assuming the register is not already assigned.
+    ///
+    /// # Errors
+    /// This method should halt if the given register is a register member.
+    /// This method should halt if the given register is an input register.
+    /// This method should halt if the register is already used.
+    fn store_literal(&mut self, stack: &impl StackTrait<N>, register: &Register<N>, literal: Literal<N>) -> Result<()> {
+        self.store(stack, register, Value::Plaintext(Plaintext::from(literal)))
+    }
+}
+
+/// This trait is intended to be implemented only by `synthesizer_process::Registers`.
+///
+/// We make it a trait only to avoid circular dependencies.
+pub trait RegistersCircuit<N: Network, A: circuit::Aleo<Network = N>> {
     /// Returns the transition signer, as a circuit.
     fn signer_circuit(&self) -> Result<circuit::Address<A>>;
 
@@ -198,51 +262,7 @@ pub trait RegistersSignerCircuit<N: Network, A: circuit::Aleo<Network = N>> {
 
     /// Sets the transition view key, as a circuit.
     fn set_tvk_circuit(&mut self, tvk_circuit: circuit::Field<A>);
-}
 
-pub trait RegistersLoad<N: Network> {
-    /// Loads the value of a given operand.
-    ///
-    /// # Errors
-    /// This method should halt if the register locator is not found.
-    /// In the case of register members, this method should halt if the member is not found.
-    fn load(&self, stack: &impl StackTrait<N>, operand: &Operand<N>) -> Result<Value<N>>;
-
-    /// Loads the literal of a given operand.
-    ///
-    /// # Errors
-    /// This method should halt if the given operand is not a literal.
-    /// This method should halt if the register locator is not found.
-    /// In the case of register members, this method should halt if the member is not found.
-    #[inline]
-    fn load_literal(&self, stack: &impl StackTrait<N>, operand: &Operand<N>) -> Result<Literal<N>> {
-        match self.load(stack, operand)? {
-            Value::Plaintext(Plaintext::Literal(literal, ..)) => Ok(literal),
-            Value::Plaintext(Plaintext::Struct(..))
-            | Value::Plaintext(Plaintext::Array(..))
-            | Value::Record(..)
-            | Value::Future(..) => {
-                bail!("Operand must be a literal")
-            }
-        }
-    }
-
-    /// Loads the plaintext of a given operand.
-    ///
-    /// # Errors
-    /// This method should halt if the given operand is not a plaintext.
-    /// This method should halt if the register locator is not found.
-    /// In the case of register members, this method should halt if the member is not found.
-    #[inline]
-    fn load_plaintext(&self, stack: &impl StackTrait<N>, operand: &Operand<N>) -> Result<Plaintext<N>> {
-        match self.load(stack, operand)? {
-            Value::Plaintext(plaintext) => Ok(plaintext),
-            Value::Record(..) | Value::Future(..) => bail!("Operand must be a plaintext"),
-        }
-    }
-}
-
-pub trait RegistersLoadCircuit<N: Network, A: circuit::Aleo<Network = N>> {
     /// Loads the value of a given operand.
     ///
     /// # Errors
@@ -256,7 +276,6 @@ pub trait RegistersLoadCircuit<N: Network, A: circuit::Aleo<Network = N>> {
     /// This method should halt if the given operand is not a literal.
     /// This method should halt if the register locator is not found.
     /// In the case of register members, this method should halt if the member is not found.
-    #[inline]
     fn load_literal_circuit(&self, stack: &impl StackTrait<N>, operand: &Operand<N>) -> Result<circuit::Literal<A>> {
         match self.load_circuit(stack, operand)? {
             circuit::Value::Plaintext(circuit::Plaintext::Literal(literal, ..)) => Ok(literal),
@@ -273,7 +292,6 @@ pub trait RegistersLoadCircuit<N: Network, A: circuit::Aleo<Network = N>> {
     /// This method should halt if the given operand is not a plaintext.
     /// This method should halt if the register locator is not found.
     /// In the case of register members, this method should halt if the member is not found.
-    #[inline]
     fn load_plaintext_circuit(
         &self,
         stack: &impl StackTrait<N>,
@@ -284,30 +302,7 @@ pub trait RegistersLoadCircuit<N: Network, A: circuit::Aleo<Network = N>> {
             circuit::Value::Record(..) | circuit::Value::Future(..) => bail!("Operand must be a plaintext"),
         }
     }
-}
 
-pub trait RegistersStore<N: Network> {
-    /// Assigns the given value to the given register, assuming the register is not already assigned.
-    ///
-    /// # Errors
-    /// This method should halt if the given register is a register member.
-    /// This method should halt if the given register is an input register.
-    /// This method should halt if the register is already used.
-    fn store(&mut self, stack: &impl StackTrait<N>, register: &Register<N>, stack_value: Value<N>) -> Result<()>;
-
-    /// Assigns the given literal to the given register, assuming the register is not already assigned.
-    ///
-    /// # Errors
-    /// This method should halt if the given register is a register member.
-    /// This method should halt if the given register is an input register.
-    /// This method should halt if the register is already used.
-    #[inline]
-    fn store_literal(&mut self, stack: &impl StackTrait<N>, register: &Register<N>, literal: Literal<N>) -> Result<()> {
-        self.store(stack, register, Value::Plaintext(Plaintext::from(literal)))
-    }
-}
-
-pub trait RegistersStoreCircuit<N: Network, A: circuit::Aleo<Network = N>> {
     /// Assigns the given value to the given register, assuming the register is not already assigned.
     ///
     /// # Errors
@@ -327,7 +322,6 @@ pub trait RegistersStoreCircuit<N: Network, A: circuit::Aleo<Network = N>> {
     /// This method should halt if the given register is a register member.
     /// This method should halt if the given register is an input register.
     /// This method should halt if the register is already used.
-    #[inline]
     fn store_literal_circuit(
         &mut self,
         stack: &impl StackTrait<N>,
