+++ synthesizer/program/tests/helpers/sample.rs
@@ -19,10 +19,7 @@ use console::{
     prelude::*,
     program::{Identifier, Literal, Plaintext, Register, Value},
 };
-use snarkvm_synthesizer_program::{
-    FinalizeGlobalState,
-    traits::{RegistersStore, RegistersStoreCircuit},
-};
+use snarkvm_synthesizer_program::{FinalizeGlobalState, RegistersCircuit as _, RegistersTrait as _};
 use synthesizer_process::{Authorization, CallStack, FinalizeRegisters, Registers, Stack};
 
 type CurrentNetwork = MainnetV0;
