+++ synthesizer/program/tests/instruction/commit.rs
@@ -34,8 +34,8 @@ use snarkvm_synthesizer_program::{
     Opcode,
     Operand,
     Program,
-    RegistersLoad,
-    RegistersLoadCircuit,
+    RegistersCircuit as _,
+    RegistersTrait as _,
 };
 use synthesizer_process::{Process, Stack};
