+++ synthesizer/program/tests/instruction/is.rs
@@ -30,8 +30,8 @@ use snarkvm_synthesizer_program::{
     Opcode,
     Operand,
     Program,
-    RegistersLoad,
-    RegistersLoadCircuit,
+    RegistersCircuit as _,
+    RegistersTrait as _,
 };
 use synthesizer_process::{Process, Stack};
